Refactor error handling
Move all handling to dedicated class.
